mainainability fixes
